chore: Extract translateFunctionTypeAnnotation into a common function
